% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_bad_argument}
\alias{abort_bad_argument}
\title{Abort bad function arguments}
\usage{
abort_bad_argument(arg, must, not = NULL)
}
\arguments{
\item{arg}{Function argument name as character vector}

\item{must}{Text to relate argument's name to its expected type}

\item{not}{Function argument's name}
}
\value{
Error handler with templated message and metadata
}
\description{
Error handler to abort a bad argument based on its actual vs. expected type
}
