% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_credentials}
\alias{clean_credentials}
\title{Clean up credentials}
\usage{
clean_credentials(x)
}
\arguments{
\item{x}{Character vector of credentials}
}
\value{
List of cleaned character vectors, with one list element per element of \code{x}
}
\description{
Clean up credentials
}
