% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_fun_wait}
\alias{do_fun_wait}
\title{Safe execution of a function that might fail}
\usage{
do_fun_wait(.f, n_tries, sleep_for = 1L, ...)
}
\arguments{
\item{.f}{A function}

\item{n_tries}{Number of times to try executing the function}

\item{sleep_for}{Number of seconds to make system sleep after each unsuccessful attempt}

\item{...}{Arguments for .f}
}
\description{
Safe execution of a function that might fail
}
\examples{
# Function success rate is `p`
get_data <- function(p = 0.8) {
  x <- rbinom(1, 1, p)
  ifelse(x == 0, "OK", stop("Error: too many calls!"))
  }

# Success on 3rd attempt
set.seed(556)
do_fun_wait(get_data, 10, 1)

# NULL if failed after n_tries
set.seed(55)
do_fun_wait(get_data, 10, 1)

# Pass .f arguments into ...
do_fun_wait(get_data, 10, 1, p = 0.6)
}
\seealso{
\url{https://www.brodrigues.co/blog/2018-03-12-keep_trying/}
}
