% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{flatten_npi}
\alias{flatten_npi}
\title{Flatten NPI search results}
\usage{
flatten_npi(df)
}
\arguments{
\item{df}{data frame containing columns named "number", "taxonomies", "addresses", and "identifiers"}
}
\value{
data frame (tibble) with flattened list columns
}
\description{
This function takes a data frame produced by `search_npi()` and returns a data fram with several list columns flattened. It left joins the data frame by `number` (NPI number) to the unnested list columns, "taxonomies", "addresses", and "identifiers". The function adds suffixes to non-key columns with identical names to avoid name clashes and identify the source of unnested columns.
}
