% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{flatten_results}
\alias{flatten_results}
\title{Flatten NPI search results}
\usage{
flatten_results(df, key)
}
\arguments{
\item{df}{data frame containing the results of a call to `search_npi()`}

\item{key}{quoted column name from \code{df} to use as a matching key}
}
\value{
data frame (tibble) with flattened list columns
}
\description{
This function takes a data frame produced by `search_npi()` and returns a data fram with several list columns flattened. It left joins the data frame by `npi` (NPI number) to the unnested list columns. The function adds suffixes to non-key columns with identical names to avoid name clashes and identify the source of unnested columns.
}
