% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_list_col}
\alias{get_list_col}
\title{Extract list column by key}
\usage{
get_list_col(df, list_col, key)
}
\arguments{
\item{df}{data frame}

\item{list_col}{list column in \code{df}}

\item{key}{key column in \code{df}}
}
\value{
data frame with \code{key} and unnested \code{list_col}
}
\description{
Extract list column by key
}
