% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_results}
\alias{get_results}
\title{get_results}
\usage{
get_results(npi_api)
}
\arguments{
\item{npi_api}{List of parsed results from npi_api()}
}
\value{
Tibble of cleaned search results
}
\description{
get_results
}
