% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_url}
\alias{get_url}
\title{Get URL}
\usage{
get_url(query = NULL, url = NULL, ua = NULL, sleep = NULL)
}
\arguments{
\item{query}{List of key-value pairs of query parameters}

\item{url}{Base URL character string to which \code{query} will be appended in a URI-compliant string}

\item{ua}{User agent character string}

\item{sleep}{Number of seconds to pause before issuing request}
}
\value{
List response object from the GET request
}
\description{
Issue a GET request to a URL with a user agent included in the header. Optionally delay the request to avoid too many requests in a short amount of time.
}
