% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hyphenate_full_zip}
\alias{hyphenate_full_zip}
\title{Add hyphen to 9-digit ZIP codes}
\usage{
hyphenate_full_zip(x)
}
\arguments{
\item{x}{Character or numeric vector containing ZIP code(s)}
}
\value{
Length \code{x} character vector hyphenated for ZIP+4
}
\description{
Add hyphen to 9-digit ZIP codes
}
