% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{luhn_check}
\alias{luhn_check}
\title{Validate Luhn check digit}
\usage{
luhn_check(number, return_logical = TRUE)
}
\arguments{
\item{number}{Number consisting of digits 0-9}

\item{return_logical}{Boolean flag to control whether the result is a logical value or remainder}
}
\value{
Boolean. If \code{return_logical} is TRUE, returns a logical value indicating whether the \code{number} validates; if FALSE, returns the remainder from the expected number
}
\description{
\code{luhn_check} validates a number based on the Luhn algorithm.
}
\references{
\url{http://scott.sherrillmix.com/blog/tag/luhn-algorithm/}
}
\seealso{
\url{https://en.wikipedia.org/wiki/Luhn_algorithm}
}
