% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npi_results_s3.R
\name{new_npi_results}
\alias{new_npi_results}
\title{Construct an \code{npi_results} S3 object}
\usage{
new_npi_results(x, ...)
}
\arguments{
\item{x}{A tibble}
}
\value{
A tibble with S3 class \code{npi_results}
}
\description{
Creates an \code{npi_results} S3 object from a tibble. See \code{\link{validate_npi_results}} for other requirements for this class.
}
\keyword{internal}
