% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{npi_api}
\alias{npi_api}
\title{National Provider Identifier API client}
\usage{
npi_api(query)
}
\arguments{
\item{query}{List of query parameters}
}
\value{
\code{npi_api} S3 class containing the API content, URL, and response.
}
\description{
API client to the U.S. National Provider Identifier (NPI) public registry.
}
