% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{npi_config}
\alias{npi_config}
\title{Specify configuration for API request}
\usage{
npi_config()
}
\description{
User can set their own user agent as a value for the
`npi_user_agent` option via options(). See \pkg{npi} for an example.
}
\keyword{internal}
