% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npi_handle_requests.R
\name{npi_handle_requests}
\alias{npi_handle_requests}
\title{Control API requests}
\usage{
npi_handle_requests(params)
}
\arguments{
\item{params}{A list of query parameters.}
}
\value{
A list of API responses.
}
\description{
Gets the maximum number of records allowed by the API in the fewest number
of requests.
}
\keyword{internal}
