% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{npi_handle_response}
\alias{npi_handle_response}
\title{Handle errors returned by the API}
\usage{
npi_handle_response(resp)
}
\arguments{
\item{resp}{Response object from a REST API request}
}
\value{
An error if present, otherwise the API response
}
\description{
Inspects the API response object's status and handle problems as follows:
  \itemize{
    \item Status codes of 400 or greater throw a \code{request_failed_error}
    \item Responses that do not return json content throw a
      \code{http_type_error}
    \item Responses containing other errors from the API throw a
      \code{request_logic_error}
  }
}
\keyword{internal}
