% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{npi_is_valid}
\alias{npi_is_valid}
\title{Check if candidate NPI number is valid}
\usage{
npi_is_valid(x)
}
\arguments{
\item{x}{10-digit candidate NPI number}
}
\value{
Boolean indicating whether \code{npi} is valid
}
\description{
Check whether a number is a valid NPI number per the specifications detailed in the Final Rule for the Standard Unique Health Identifier for Health Care Providers (69 FR 3434).
}
\examples{
npi_is_valid(1234567893)
npi_is_valid(1234567898)
}
\references{
\url{http://scott.sherrillmix.com/blog/tag/luhn-algorithm/}
}
\seealso{
\url{https://www.cms.gov/Regulations-and-Guidance/Administrative-Simplification/NationalProvIdentStand/Downloads/NPIcheckdigit.pdf}
}
