% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npi_results_s3.R
\name{npi_summarize.npi_results}
\alias{npi_summarize.npi_results}
\title{Summary method for \code{npi_results} S3 object}
\usage{
\method{npi_summarize}{npi_results}(object, ...)
}
\arguments{
\item{object}{An \code{npi_results} S3 object}

\item{...}{Additional optional arguments}
}
\value{
Tibble containing the following columns:
  \describe{
    \item{\code{npi}}{National Provider Identifier (NPI) number}
    \item{\code{name}}{Provider's first and last name for individual
    providers, organization name for organizational providers.}
    \item{\code{enumeration_type}}{Type of provider associated with the NPI,
      either "Individual" or "Organizational"}
    \item{\code{primary_practice_address}}{Full address of the provider's
      primary practice location}
    \item{\code{phone}}{Provider's telephone number}
    \item{\code{primary_taxonomy}}{Primary taxonomy description. If no
      taxonomy is marked as primary for a record, the first listed taxonomy
      is used.}
  }
}
\description{
Print a human-readable overview of each record return in the results from a
call to \code{\link{npi_search}}. The format of the summary is modeled after
the one offered on the NPI registry website.
}
\details{
Missing optional second address lines are treated as empty strings.
  Missing required address components leave the full address missing.
}
\examples{
data(npis)
npi_summarize(npis)
}
