% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{npi_url}
\alias{npi_url}
\title{Construct request URL using BASE_URL}
\usage{
npi_url()
}
\value{
The value of BASE_URL set internally in \code{R/utils.R}
}
\description{
Construct request URL using BASE_URL
}
\keyword{internal}
