% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{prep_list_col}
\alias{prep_list_col}
\title{Package npi_api entries as list of cleaned dfs}
\usage{
prep_list_col(npi_api, name)
}
\arguments{
\item{npi_api}{Search result from search_npi}

\item{name}{Name of entity in search_npi result}
}
\value{
List of tibbles corresponding to named element
}
\description{
Cleaning includes turning all empty character strings to NAs
}
