% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.npi_api}
\alias{print.npi_api}
\title{Print method for npi_api S3 class}
\usage{
\method{print}{npi_api}(x, ...)
}
\arguments{
\item{x}{npi_api S3 class object}

\item{...}{Optional arguments}
}
\description{
Print the structure of the content in an \code{npi_api} S3 class object.
}
