% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{summary.npi_results}
\alias{summary.npi_results}
\title{Summary method for npi_results S3 objects}
\usage{
\method{summary}{npi_results}(object, ...)
}
\arguments{
\item{object}{`npi_results` S3 object}

\item{...}{Additional aptional arguments}
}
\value{
Tibble containing the following columns: `npi`, `name`, `provider_type`, `primary_practice_address`, `phone`, and `primary_taxonomy`.
}
\description{
Modeled after summary profile presented on the NPPES registry
website.
}
