% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{validate_response}
\alias{validate_response}
\title{Handle errors from API request}
\usage{
validate_response(resp)
}
\arguments{
\item{resp}{Response object from a REST API request}
}
\description{
Inspect API response object's status and handle business logic errors returned by API.
}
