% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npi_results_s3.R
\name{flatten_npi}
\alias{flatten_npi}
\title{S3 method to flatten an \code{npi_results} object}
\usage{
flatten_npi(df, cols, key)
}
\arguments{
\item{df}{A data frame containing the results of a call to
\code{\link{search_npi}}.}

\item{cols}{If non-NULL, only the named columns specified here will be be
flattened and returned along with \code{npi}.}

\item{key}{A quoted column name from \code{df} to use as a matching key. The
default value is \code{"npi"}.}
}
\description{
S3 method to flatten an \code{npi_results} object
}
