% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_list_col}
\alias{get_list_col}
\title{Extract list column by key}
\usage{
get_list_col(df, list_col = NULL, key = "npi")
}
\arguments{
\item{df}{A data frame}

\item{list_col}{The quoted name of a list column in \code{df}}

\item{key}{One or more quoted names of columns in \code{df} to keep
 alongside the unnested columns of \code{list_col} in the result.
If the value is \code{NULL}, the result will just unnest \code{list_col}.}
}
\value{
A data frame with the column(s) specified in \code{key}
  followed by the columns unnested from \code{list_col}
}
\description{
Extract list column by key
}
\examples{
# Load sample data
data("npis")

# Get basic list column by NPI
get_list_col(npis, "basic")
get_list_col(npis, "taxonomies")
}
