% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_results}
\alias{get_results}
\title{Get results}
\usage{
get_results(results)
}
\arguments{
\item{results}{Results object from \code{\link{npi_search}}}
}
\description{
Get results
}
\keyword{internal}
