% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hyphenate_full_zip}
\alias{hyphenate_full_zip}
\title{Format United States (US) ZIP codes}
\usage{
hyphenate_full_zip(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Length \code{x} character vector hyphenated for ZIP+4 or 5-digit ZIP. Invalid elements of \code{x} are not formatted.
}
\description{
Format United States (US) ZIP codes
}
