% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{npi_api}
\alias{npi_api}
\title{Make generic API request and handle the response}
\usage{
npi_api(verb, url, config = list(), ...)
}
\arguments{
\item{verb}{httr request verb such as GET, POST, etc.}

\item{url}{URL to which the request will be sent}

\item{config}{List of key-value configuration parameters}

\item{...}{Optional arguments to be passed to `httr::modify_url()`}
}
\value{
Response object or error
}
\description{
Make generic API request and handle the response
}
\keyword{internal}
