% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npi_results_s3.R
\name{npi_summarize}
\alias{npi_summarize}
\title{S3 method to summarize an \code{npi_results} object}
\usage{
npi_summarize(object, ...)
}
\arguments{
\item{object}{An \code{npi_results} S3 object}

\item{...}{Additional optional arguments}
}
\description{
S3 method to summarize an \code{npi_results} object
}
