% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{search_npi}
\alias{search_npi}
\title{Search the NPI Registry}
\usage{
search_npi(npi = NULL, provider_type = NULL, taxonomy = NULL,
  first_name = NULL, last_name = NULL, org_name = NULL,
  address_purpose = NULL, city = NULL, state = NULL, postal_code = NULL,
  country_code = NULL, limit = 200, skip = NULL)
}
\arguments{
\item{npi}{10-digit National Provider Identifier number assigned to the provider.}

\item{provider_type}{The API can be refined to retrieve only Individual Providers or Organizational Providers. When it is not specified, both Type 1 and Type 2 NPIs will be returned. When using the Enumeration Type, it cannot be the only criteria entered. Additional criteria must also be entered as well. Valid values are: NPI-1: Individual Providers (Type 1) NPIs; NPI-2: Organizational Providers (Type 2) NPIs}

\item{taxonomy}{Search for providers by their taxonomy by entering the taxonomy description.}

\item{first_name}{This field only applies to Individual Providers. Trailing wildcard entries are permitted requiring at least two characters to be entered (e.g. "jo*" ). This field allows the following special characters: ampersand, apostrophe, colon, comma, forward slash, hyphen, left and right parentheses, period, pound sign, quotation mark, and semi-colon.}

\item{last_name}{This field only applies to Individual Providers. Trailing wildcard entries are permitted requiring at least two characters to be entered. This field allows the following special characters: ampersand, apostrophe, colon, comma, forward slash, hyphen, left and right parentheses, period, pound sign, quotation mark, and semi-colon.}

\item{org_name}{This field only applies to Organizational Providers. Trailing wildcard entries are permitted requiring at least two characters to be entered. This field allows the following special characters: ampersand, apostrophe,"at" sign, colon, comma, forward slash, hyphen, left and right parentheses, period, pound sign, quotation mark, and semi-colon. Both the Organization Name and Other Organization Name fields associated with an NPI are examined for matching contents, therefore, the results might contain an organization name different from the one entered in the Organization Name criterion.}

\item{address_purpose}{Refers to whether the address information entered pertains to the provider's Mailing Address or the provider's Practice Location Address. When not specified, the results will contain the providers where either the Mailing Address or the Practice Location Addresses match the entered address information. Valid values are: "LOCATION", "MAILING"}

\item{city}{The City associated with the provider's address identified in Address Purpose. To search for a Military Address enter either APO or FPO into the City field. This field allows the following special characters: ampersand, apostrophe, colon, comma, forward slash, hyphen, left and right parentheses, period, pound sign, quotation mark, and semi-colon.}

\item{state}{The State abbreviation associated with the provider's address identified in Address Purpose. This field cannot be used as the only input criterion. If this field is used, at least one other field, besides the Enumeration Type and Country, must be populated. Valid values for states: https://npiregistry.cms.hhs.gov/registry/API-State-Abbr}

\item{postal_code}{The Postal Code associated with the provider's address identified in Address Purpose. There is an implied trailing wildcard. If you enter a 5 digit postal code, it will match any appropriate 9 digit (zip+4) codes in the data.}

\item{country_code}{The Country associated with the provider's address identified in Address Purpose. This field can be used as the only input criterion as long as the value selected is not US (United States). Valid values for country codes: https://npiregistry.cms.hhs.gov/registry/API-Country-Abbr}

\item{limit}{Limit the results returned. The default value is 10; however, the value can be set to any value from 1 to 200.}

\item{skip}{The first N (value entered) results meeting the entered criteria will be bypassed and will not be included in the output.}
}
\value{
\code{npi_api} S3 class containing the API content, URL, and response.
}
\description{
Wrapper function to search the U.S. National Provider Identifier (NPI) Registry using search parameters exposed by the registry's API.
}
